fix(config): decouple ptc_agent config from src package imports
